# EventManager/app/docs/swagger.yaml
basePath: /api/event-manager
definitions:
  hateoas.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
      title:
        type: string
    type: object
  httpdto.HttpCreateEvent:
    properties:
      description:
        type: string
      id_owner:
        type: integer
      location:
        type: string
      name:
        type: string
      seats:
        type: integer
    required:
    - id_owner
    - name
    type: object
...
  httpdto.HttpResponseEvent:
    properties:
      event:
        $ref: '#/definitions/httpdto.httpResponseEvent'
    type: object
...
  httpdto.HttpUpdateEvent:
    properties:
      description:
        type: string
      id_owner:
        type: integer
      location:
        type: string
      name:
        type: string
      seats:
        type: integer
    type: object
...
  httpdto.httpResponseEvent:
    properties:
      _links:
        additionalProperties:
          $ref: '#/definitions/hateoas.Link'
        type: object
      description:
        type: string
      id:
        type: integer
      id_owner:
        type: integer
      location:
        type: string
      name:
        type: string
      seats:
        type: integer
    type: object
...
host: localhost:12345
info:
  contact: {}
  description: Event management service for the POS system. Handles events, event
    packets, tickets, and inclusions.
  title: EventManager Service API
  version: "1.0"
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: Get a paginated list of events with optional filters
      parameters:
      - description: Bearer token (optional)
        in: header
        name: Authorization
        type: string
      - description: Filter by event name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by location (partial match)
        in: query
        name: location
        type: string
      - description: Filter by description (partial match)
        in: query
        name: description
        type: string
      - description: Minimum number of seats
        in: query
        name: min_seats
        type: integer
      - description: Maximum number of seats
        in: query
        name: max_seats
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: per_page
        type: integer
      - description: Sort field (e.g., 'name', 'created_at')
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of events
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid query parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List and filter events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event with the provided details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event details
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpCreateEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Event created successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEvent'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict - event already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new event
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event by its ID (only the owner can delete)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEvent'
        "400":
          description: Invalid event ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - not the event owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Retrieve a specific event by its unique identifier
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token (optional)
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event details
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEvent'
        "400":
          description: Invalid event ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event by ID
      tags:
      - events
    patch:
      consumes:
      - application/json
      description: Partially update event details (PATCH - only provided fields are
        updated)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpUpdateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: Event updated successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEvent'
        "400":
          description: Invalid request body or event ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - not the event owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing event
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"